+++ docs/changelog.md
@@ -2,6 +2,12 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+# [0.4.0] - 2024-09-20
+## Added
+- Standardization of peps using bedbase bedms schema
+- Reference validator module
+
+
 # [0.4.0] - 2024-08-26
 ## Added
 - Added bbuploader (GEO uploader)